VTX-3411: move some of the code to executor replication metrics
