+++ ballista/core/Cargo.toml
@@ -47,7 +47,6 @@ simd = ["datafusion/simd"]
 ahash = { version = "0.8", default-features = false }
 arrow-flight = { workspace = true }
 async-trait = "0.1.73"
-bytes = "1.0"
 chrono = { version = "0.4", default-features = false }
 clap = { version = "3", features = ["derive", "cargo"] }
 datafusion = { workspace = true }
@@ -72,7 +71,6 @@ sqlparser = { workspace = true }
 sys-info = "0.9.0"
 tokio = "1.0"
 tokio-stream = { version = "0.1", features = ["net"] }
-tokio-util = { version = "0.7.10", features = ["compat"] }
 tonic = { workspace = true }
 tracing = "0.1.36"
 url = "2.2"
