+++ ballista/core/src/serde/scheduler/mod.rs
@@ -124,6 +124,10 @@ impl PartitionStats {
         }
     }
 
+    pub fn num_batches(&self) -> Option<u64> {
+        self.num_batches
+    }
+
     pub fn arrow_struct_repr(self) -> Field {
         Field::new(
             "partition_stats",
