+++ ballista/executor/Cargo.toml
@@ -42,6 +42,7 @@ arrow = { workspace = true }
 arrow-flight = { workspace = true }
 async-trait = "0.1.73"
 ballista-core = { path = "../core", version = "0.11.0", features = ["s3"] }
+bytes = "1.0"
 chrono = { version = "0.4", default-features = false }
 configure_me = { workspace = true }
 dashmap = "5.4.0"
@@ -66,6 +67,7 @@ tokio = { version = "1.0", features = [
     "signal",
 ] }
 tokio-stream = { version = "0.1", features = ["net"] }
+tokio-util = { version = "0.7.10", features = ["compat"] }
 tonic = { workspace = true }
 tracing = "0.1.36"
 tracing-appender = "0.2.2"
